catalog: twelve MLX models across coding, writing, science, uncensored, general, and embedding
The catalog moves to lib/stdlib/modelCatalog.ts, since localModels.ts
was at the line limit. Three new categories: writing, science,
uncensored. Sizes and licenses were read from the Hub API on 2026-09-08.
