+++ packages/agency-lang/data/model-catalog.json
@@ -264,6 +264,126 @@
       "license": "apache-2.0",
       "description": "Returns 768-dim embeddings. Pair with a chat model for RAG.",
       "sha256": "d4e388894e09cf3816e8b0896d81d265b55e7a9fff9ab03fe8bf4ef5e11295ac"
+    },
+    "deepseek-v4-flash-mlx": {
+      "backend": "mlx",
+      "uri": "mlx:mlx-community/DeepSeek-V4-Flash-4bit",
+      "params": "284B (A13B)",
+      "sizeBytes": 151500000000,
+      "category": "coding",
+      "contextWindow": 1048576,
+      "license": "mit",
+      "description": "DeepSeek's sparse coder, 79% on SWE-bench Verified. Needs about 160 GB of memory."
+    },
+    "qwen3-coder-next-mlx": {
+      "backend": "mlx",
+      "uri": "mlx:mlx-community/Qwen3-Coder-Next-4bit",
+      "params": "80B (A3B)",
+      "sizeBytes": 44900000000,
+      "category": "coding",
+      "contextWindow": 262144,
+      "license": "apache-2.0",
+      "description": "Qwen's agentic coder with 3B active. Fast enough to be a daily driver on 64 GB."
+    },
+    "qwen3-coder-30b-a3b-mlx": {
+      "backend": "mlx",
+      "uri": "mlx:mlx-community/Qwen3-Coder-30B-A3B-Instruct-4bit",
+      "params": "30B (A3B)",
+      "sizeBytes": 17200000000,
+      "category": "coding",
+      "contextWindow": 262144,
+      "license": "apache-2.0",
+      "description": "The MLX build of qwen3-coder-30b-a3b. Small, fast agentic coder."
+    },
+    "qwen3-235b-a22b-2507-mlx": {
+      "backend": "mlx",
+      "uri": "mlx:mlx-community/Qwen3-235B-A22B-Instruct-2507-4bit",
+      "params": "235B (A22B)",
+      "sizeBytes": 132300000000,
+      "category": "writing",
+      "contextWindow": 262144,
+      "license": "apache-2.0",
+      "description": "Top open-weight model on the EQ-Bench creative writing leaderboard. Needs about 140 GB of memory."
+    },
+    "gemma-4-31b-mlx": {
+      "backend": "mlx",
+      "uri": "mlx:mlx-community/gemma-4-31b-it-4bit",
+      "params": "31B",
+      "sizeBytes": 18400000000,
+      "category": "writing",
+      "contextWindow": 262144,
+      "license": "apache-2.0",
+      "description": "Google's dense 31B. The best prose writer at its size."
+    },
+    "qwen3.8-27b-mlx": {
+      "backend": "mlx",
+      "uri": "mlx:mlx-community/Qwen3.8-27B-4bit",
+      "params": "27B",
+      "sizeBytes": 16100000000,
+      "category": "science",
+      "contextWindow": 262144,
+      "license": "apache-2.0",
+      "description": "Alibaba's current 27B, 89 on GPQA Diamond. Multimodal. The research and science default."
+    },
+    "gpt-oss-120b-mlx": {
+      "backend": "mlx",
+      "uri": "mlx:mlx-community/gpt-oss-120b-MXFP4-Q8",
+      "params": "120B (A5B)",
+      "sizeBytes": 63400000000,
+      "category": "science",
+      "contextWindow": 131072,
+      "license": "apache-2.0",
+      "description": "OpenAI's open reasoning model with adjustable effort. Strong on math and science, weaker on code."
+    },
+    "qwen3.6-40b-deckard-uncensored-mlx": {
+      "backend": "mlx",
+      "uri": "mlx:mlx-community/Qwen3.6-40B-Claude-4.6-Opus-Deckard-Heretic-Uncensored-Thinking-8bit",
+      "params": "40B",
+      "sizeBytes": 41500000000,
+      "category": "uncensored",
+      "contextWindow": 262144,
+      "license": "apache-2.0",
+      "description": "Heretic-uncensored Qwen3.6, expanded to 40B and further trained. Beats the base 27B on 6 of 7 benchmarks."
+    },
+    "qwen3.8-27b-uncensored-mlx": {
+      "backend": "mlx",
+      "uri": "mlx:mlx-community/Qwen3.8-27B-Uncensored-OptiQ-4bit",
+      "params": "27B",
+      "sizeBytes": 19800000000,
+      "category": "uncensored",
+      "contextWindow": 262144,
+      "license": "apache-2.0",
+      "description": "Qwen3.8-27B with refusals removed. Text-only."
+    },
+    "gemma-4-31b-uncensored-mlx": {
+      "backend": "mlx",
+      "uri": "mlx:mlx-community/gemma-4-31B-it-uncensored-heretic-4bit",
+      "params": "31B",
+      "sizeBytes": 17300000000,
+      "category": "uncensored",
+      "contextWindow": 262144,
+      "license": "apache-2.0",
+      "description": "Gemma 4 31B with refusals removed by Heretic. Text-only."
+    },
+    "qwen3.5-4b-mlx": {
+      "backend": "mlx",
+      "uri": "mlx:mlx-community/Qwen3.5-4B-MLX-4bit",
+      "params": "4B",
+      "sizeBytes": 3060000000,
+      "category": "general",
+      "contextWindow": 262144,
+      "license": "apache-2.0",
+      "description": "The MLX build of qwen3.5-4b. Small general model for a laptop."
+    },
+    "qwen3-embedding-4b-mlx": {
+      "backend": "mlx",
+      "uri": "mlx:mlx-community/Qwen3-Embedding-4B-4bit-DWQ",
+      "params": "4B",
+      "sizeBytes": 2280000000,
+      "category": "embedding",
+      "contextWindow": 40960,
+      "license": "apache-2.0",
+      "description": "Returns 2560-dim embeddings. mlx_lm.server has no embeddings route, so serve it with another MLX server."
     }
   }
 }
